fix btap

+++ ss11_mvc_model/exercise/product-management/src/main/java/com/example/productmanagement/service/impl/ProductServiceImpl.java
@@ -4,6 +4,7 @@ import com.example.productmanagement.model.Product;
 import com.example.productmanagement.repository.IProductRepo;
 import com.example.productmanagement.repository.impl.ProductRepoImpl;
 import com.example.productmanagement.service.IProductService;
+import org.omg.CosNaming.NamingContextPackage.NotFound;
 
 import java.util.List;
 
@@ -30,7 +31,12 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public void remove(int id) {
-        productRepo.remove(id);
+    public void remove(int id) throws Exception {
+        Product product = productRepo.findById(id);
+        if (product == null) {
+           throw new Exception("Khong tim thay");
+        } else {
+            productRepo.remove(id);
+        }
     }
 }
